M-07: her istemci kendi alanında
İki ajan aynı loopback portunda iki yabancıdır. Bugüne kadar öyle
davranılmıyordu: tek bir tutamak defteri ve sahipsiz öneriler vardı.

Tutamak defteri artık istemci başına (`ai::HandleScopes`). Bu kozmetik bir
ayrım değildi: `HandleStore::next_id` bir sayaç hashliyor, yani tek defterde
bile olsalar iki istemcinin ikinci tutamağı aynı kimliği taşırdı — bir ajan
hiç okumadığı geometrinin tutamağını tahmin etmeden eline geçirebilirdi.
`next_id`'nin kendi yorumu ("her oturumun kendi sayacı") ancak şimdi doğru.

Önerilerin sahibi var (`PlanStore::owned_by`, `find_for`, `append_for`). Asıl
madde okumak değil GENİŞLETMEK: öneri kimliği istemcinin cevabında geçtiği
için gizli değil, ve adım ekleyebilen ikinci bir istemci, mühendisin kartta
okumadığı bir satırı onun onayıyla uygulatırdı; denetim kaydı da o adım için
yanlış istemciyi gösterirdi. "Sizin değil" ile "yok" aynı cevabı alıyor —
ayrı bir ret, bir istemciye bir başkasının planını doğrulayan kâhin olurdu.

Boş istemci adı masadaki kişidir ve kapsamsızdır: öneriyi uygulayan o, öneri
panosu da kimin açtığına bakmadan hepsini listeliyor.

Kapsamlama Qt'siz `/src/ai`'de, böylece testler gerçek kodu sınıyor; sahte
dispatcher da aynı iki tipi kullanıyor. `AiService::propose` adım ekleme
sözleşmesini artık gerçekten uyguluyor — `mcp.cpp` bunu anlatıyordu ama
uygulama her çağrıda yeni öneri açıyordu.

Yol üstünde iki clang-tidy hatası: `provider.cpp`'de iki özdeş dal tek
etikete indi, `layout_render.cpp`'de `size_t` → `int32_t` daralması açıkça
yazıldı.

make check yeşil; 798 doctest, 46/46 ctest.

+++ src/app/src/layout_render.cpp
@@ -632,7 +632,7 @@ void paint_layout_page(QPainter& painter, const QRectF& target, const core::Docu
                                  item->text,
                                  [&] {
                                      core::SheetContext ctx = sheet_context(layout, owner, facts);
-                                     ctx.page               = index + 1;
+                                     ctx.page               = static_cast<std::int32_t>(index) + 1;
                                      return ctx;
                                  }(),
                                  trouble)));
